Added SeedDataService to populate data into de MemoryDB for test
