+++ WorkWhiz.API/Program.cs
@@ -1,13 +1,13 @@
 using Microsoft.EntityFrameworkCore;
 using System;
 using WorkWhiz.Infraestructure;
+using WorkWhiz.Infraestructure.Interfaces;
+using WorkWhiz.Infraestructure.Repositories;
 
 var builder = WebApplication.CreateBuilder(args);
 
 // Add services to the container.
-
 builder.Services.AddControllers();
-// Learn more about configuring Swagger/OpenAPI at https://aka.ms/aspnetcore/swashbuckle
 builder.Services.AddEndpointsApiExplorer();
 builder.Services.AddSwaggerGen();
 
@@ -16,13 +16,21 @@ builder.Services.AddDbContext<ApiContext>(options =>
     options.UseInMemoryDatabase("WorkWhizDb");
 });
 
-// Register RequestDelegate
-builder.Services.AddSingleton<RequestDelegate>(next => context => Task.CompletedTask);
 
-// Register seeding service or middleware (replace with appropriate type)
-builder.Services.AddTransient<SeedDataMiddleware>();
+// Register seeding service
+builder.Services.AddTransient<SeedDataService>();
+
+// Register other services
+builder.Services.AddScoped<IPosterRepository, PosterRepository>();
 
 var app = builder.Build();
+
+// Run seeding logic at startup
+using (var scope = app.Services.CreateScope())
+{
+    var seedService = scope.ServiceProvider.GetRequiredService<SeedDataService>();
+    await seedService.SeedAsync();  
+}
 
 // Configure the HTTP request pipeline.
 if (app.Environment.IsDevelopment())
